+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Total cells column no longer required for vi-cell XR
 - Ignore invalid first row when present for vi-cell XR files
+- Capture concentration on Nanodrop Eight files that do not have NA Type column
 - Removed hardcoding of date parsing around Gen5 plate numbers
 ### Changed
 - Corrections to the spectrophotometry/BENCHLING/2023/12 schema to account for feedback from Allotrope Modeling Working Group